requirements.txt
streamlit
sentence-transformers
faiss-cpu
duckduckgo-search
numpy
pandas
